requirements.txt
pandas
pyarrow
apache-airflow-providers-microsoft-azure